refactor(util): 重构下载文件方法并优化 Context 扩展函数
- 将下载文件方法从 StringUtil 移动到 Extensions.kt 中
- 优化 showToast 方法，使用 QVTApplication.instance 作为上下文
- 移除冗余的 Context 扩展函数
- 调整相关 Activity 和 Fragment 中的方法调用
