+++ app/src/main/java/com/xiaoniu/qqversionlist/ui/ExpUrlListAdapter.kt
@@ -28,7 +28,7 @@ import com.xiaoniu.qqversionlist.R
 import com.xiaoniu.qqversionlist.databinding.ExpLinkNextButtonBinding
 import com.xiaoniu.qqversionlist.databinding.ItemExpBackUrlCardBinding
 import com.xiaoniu.qqversionlist.util.ClipboardUtil.copyText
-import com.xiaoniu.qqversionlist.util.StringUtil.downloadFile
+import com.xiaoniu.qqversionlist.util.Extensions.downloadFile
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -106,45 +106,6 @@ class ExpUrlListAdapter(private val urlList: List<String>) :
                                         val shareIntent = Intent(Intent.ACTION_SEND).apply {
                                             type = "text/plain"
                                             putExtra(Intent.EXTRA_TEXT, url)
-                                            /*if (url.contains("downv6.qq.com")) {
-                                                if (appSize != "" && appSize != "-1" && appSize != "0") "Android QQ（${
-                                                    url.substringAfterLast(
-                                                        '/'
-                                                    )
-                                                }）（${itemView.context.getString(R.string.fileSize)}$appSize MB）\n\n${
-                                                    itemView.context.getString(
-                                                        R.string.downloadLink
-                                                    )
-                                                }$url\n\n此下载地址指向的 QQ 安装包可能属于测试版本。测试版本可能存在不可预知的稳定性问题，请明确并确保自身具备足够的风险识别和承受能力。"
-                                                else "Android QQ（${url.substringAfterLast('/')}）\n\n${
-                                                    itemView.context.getString(
-                                                        R.string.downloadLink
-                                                    )
-                                                }$url\n\n此下载地址指向的 QQ 安装包可能属于测试版本。测试版本可能存在不可预知的稳定性问题，请明确并确保自身具备足够的风险识别和承受能力。"
-                                            } else if (url.contains("imtt.dd.qq.com")) {
-                                                val appName = when {
-                                                    url.contains("com.tencent.mobileqq") -> "Android QQ"
-                                                    url.contains("com.tencent.tim") -> "Android TIM"
-                                                    url.contains("com.tencent.mm") -> "Android 微信"
-                                                    url.contains("com.tencent.wework") -> "Android 企业微信"
-                                                    url.contains("com.tencent.wetype") -> "Android 微信输入法"
-                                                    else -> url.substringAfterLast('/')
-                                                }
-
-                                                if (appSize != "" && appSize != "-1" && appSize != "0") "${appName}（大小：$appSize MB）\n\n下载地址：$url\n\n来自腾讯应用宝"
-                                                else "${appName}\n\n${
-                                                    itemView.context.getString(R.string.downloadLink)
-                                                }$url\n\n来自腾讯应用宝"
-                                            } else {
-                                                if (appSize != "" && appSize != "-1" && appSize != "0") "Android QQ（${
-                                                    url.substringAfterLast('/')
-                                                }）（大小：$appSize MB）\n\n下载地址：$url\n\n此下载地址由 TDS 腾讯端服务 Shiply 发布平台提供，指向的 QQ 安装包可能属于测试版本。测试版本可能存在不可预知的稳定性问题，请明确并确保自身具备足够的风险识别和承受能力。"
-                                                else "Android QQ（${url.substringAfterLast('/')}）\n\n${
-                                                    itemView.context.getString(
-                                                        R.string.downloadLink
-                                                    )
-                                                }$url\n\n此下载地址由 TDS 腾讯端服务 Shiply 发布平台提供，指向的 QQ 安装包可能属于测试版本。测试版本可能存在不可预知的稳定性问题，请明确并确保自身具备足够的风险识别和承受能力。"
-                                            }*/
                                         }
                                         itemView.context.startActivity(
                                             Intent.createChooser(
