+++ app/src/main/java/com/xiaoniu/qqversionlist/ui/TIMVersionAdapter.kt
@@ -141,7 +141,7 @@ class TIMVersionAdapter :
     private fun longPressCard(bindingAdapterPosition: Int, it: View) {
         if (DataStoreUtil.getBooleanKV("longPressCard", true)) showDialog(
             it.context, currentList[bindingAdapterPosition].jsonString.toPrettyFormat()
-        ) else it.context.showToast(R.string.longPressToViewSourceDetailsIsDisabled)
+        ) else showToast(R.string.longPressToViewSourceDetailsIsDisabled)
     }
 
     private fun bindDisplayInstall(
