+++ app/src/main/java/com/xiaoniu/qqversionlist/ui/TIMVersionListFragment.kt
@@ -30,9 +30,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.xiaoniu.qqversionlist.databinding.RecycleTimVersionBinding
-import com.xiaoniu.qqversionlist.util.Extensions.dp
 import com.xiaoniu.qqversionlist.util.Extensions.pxToDp
-import me.zhanghai.android.fastscroll.FastScrollerBuilder
 
 class TIMVersionListFragment : Fragment() {
     private var _fragmentBinding: RecycleTimVersionBinding? = null
