+++ app/src/test/java/com/xiaoniu/qqversionlist/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.xiaoniu.qqversionlist
 
-import org.junit.Test
-
 import org.junit.Assert.*
+import org.junit.Test
 
 /**
  * Example local unit test, which will execute on the development machine (host).
